+++ lib/src/di_module/base_di_module.dart
@@ -5,15 +5,19 @@ import 'dart:async';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:take_it/src/di_container/di_container.dart';
+import 'package:take_it/src/di_module/empty_di_module.dart';
 import 'package:take_it/take_it.dart';
 
 part '../di_scope_builder.dart';
+
+part '../di_scope_root.dart';
 
 part 'di_module.dart';
 
 part 'di_module_async.dart';
 
-/// Base class for managing Dependency Injection (DI) modules in the application.
+/// Base class for managing Dependency Injection (DI) modules
+/// in the application.
 /// It provides the core logic for setting up, resetting, and managing scopes of
 /// dependency registrations.
 ///
@@ -26,7 +30,8 @@ sealed class BaseDiModule extends ChangeNotifier implements Scope {
 
   /// Asynchronously initializes the DI module.
   ///
-  /// Depending on whether the module is synchronous or asynchronous ([DiModule] or [DiModuleAsync]),
+  /// Depending on whether the module is synchronous
+  /// or asynchronous ([DiModule] or [DiModuleAsync]),
   /// it sets up the internal [DiContainer] and marks the module as initialized.
   /// Returns a [Future] that completes once the initialization process is done.
   /// You can use it for initialize with native app splash screen
@@ -36,7 +41,8 @@ sealed class BaseDiModule extends ChangeNotifier implements Scope {
     return completer.future;
   }
 
-  /// Initializes the DI module by setting up the internal [DiContainer] and calling [onInit]
+  /// Initializes the DI module
+  /// by setting up the internal [DiContainer] and calling [onInit]
   /// after the setup process is complete.
   ///
   /// - [onInit]: Callback to be executed when initialization completes.
@@ -87,14 +93,16 @@ sealed class BaseDiModule extends ChangeNotifier implements Scope {
     _diContainer.updateScope(parent?._diContainer);
   }
 
-  /// Pops the current scope from the stack and resets the internal [DiContainer].
+  /// Pops the current scope
+  /// from the stack and resets the internal [DiContainer].
   ///
   /// Returns a [Future] that completes once the scope has been reset.
   Future<void> _popScope() async {
     await _reset(isNeedNotify: false);
   }
 
-  /// Resets the current module's scope and optionally notifies listeners about the change.
+  /// Resets the current module's scope
+  /// and optionally notifies listeners about the change.
   ///
   /// - [isNeedNotify]: If `true`, notifies listeners after the reset.
   Future<void> _reset({bool isNeedNotify = true}) async {