+++ lib/src/registrar/sync_registrar.dart
@@ -5,7 +5,8 @@ import 'dart:async';
 ///
 /// Attempting to register a type that is already registered within the
 /// current scope will throw an exception. However, if the type is registered
-/// in a parent scope, the new registration will shadow the parent's registration
+/// in a parent scope, the new registration
+/// will shadow the parent's registration
 /// for the current and child scopes.
 abstract interface class SyncRegistrar {
   /// Registers a factory for creating objects of type [T].
@@ -22,15 +23,16 @@ abstract interface class SyncRegistrar {
   /// a new instance of [T] based on this parameter.
   ///
   /// This method is useful when retrieving objects with a parameter by using
-  /// `scope.get(param)`, allowing you to pass a dynamic parameter to the factory
-  /// at runtime.
+  /// `scope.get(param)`, allowing you to pass
+  /// a dynamic parameter to the factory at runtime.
   ///
   /// Throws an exception if [T] is already registered in the current scope.
   void registerFactoryParam<T extends Object, P>(CreateFuncParam<T, P> create);
 
   /// Registers an existing instance of type [T] as a singleton.
   ///
-  /// The [instance] will always be returned for any subsequent requests for [T].
+  /// The [instance] will always be returned
+  /// for any subsequent requests for [T].
   /// Optionally, a [dispose] function can be provided to clean up resources
   /// when the scope is disposed.
   ///
@@ -43,7 +45,8 @@ abstract interface class SyncRegistrar {
   /// Registers a factory that lazily creates a singleton of type [T].
   ///
   /// The [create] function will be called only once, upon the first request
-  /// for [T], and the created instance will be reused for all subsequent requests.
+  /// for [T], and the created instance will be reused
+  /// for all subsequent requests.
   /// Optionally, a [dispose] function can be provided for resource cleanup.
   ///
   /// Throws an exception if [T] is already registered in the current scope.
@@ -56,7 +59,8 @@ abstract interface class SyncRegistrar {
 /// Signature for a function that creates instances of type [T].
 typedef CreateFunc<T> = T Function();
 
-/// Signature for a function that creates instances of type [T] with a parameter of type [P].
+/// Signature for a function that creates instances of type [T]
+/// with a parameter of type [P].
 typedef CreateFuncParam<T, P> = T Function(P param);
 
 /// Signature for a function that handles the disposal of instances of type [T].